refactor: 통합테스트 통과를 위한 코드 리팩토링
- ConcertBookingFacadeTest 의 좌석예약 기능 통합테스트를 통과하기 위해서 필요한 내용 수정
- 누락 및 버그 수정
