+++ src/main/java/io/hhplus/conbook/application/concert/ConcertBookingFacade.java
@@ -27,7 +27,7 @@ public class ConcertBookingFacade {
      * - 좌석을 미리 안 만들 경우 어떤 식으로 처리할 지는 추후 여유가 있을 때 구현
      *  (좌석의 좌표값(x,y)만 제공될 경우)
      */
-    @CacheEvict(value = "concertSchedules", key = "#serach.concertId")
+    @CacheEvict(value = "concertSchedules", key = "#booking.concertId")
     @Transactional
     public ConcertBookingResult.BookingSeat bookConcertSeat(ConcertBookingCommand.BookingSeat booking) {
         User user = userService.getUserByUUID(booking.userUUID());
