+++ src/main/java/io/hhplus/conbook/application/event/ConcertBookingEvent.java
@@ -3,12 +3,14 @@ package io.hhplus.conbook.application.event;
 import io.hhplus.conbook.domain.booking.Booking;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
+import lombok.ToString;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 @AllArgsConstructor
 @Getter
+@ToString
 public class ConcertBookingEvent {
     private Long bookingId;
     private Long concertId;
