+++ src/main/java/io/hhplus/conbook/util/JsonUtil.java
@@ -2,6 +2,7 @@ package io.hhplus.conbook.util;
 
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
@@ -9,6 +10,7 @@ public class JsonUtil {
 
     public static <T> String toJson(T obj) {
         ObjectMapper mapper = new ObjectMapper();
+        mapper.registerModule(new JavaTimeModule());
         try {
             return mapper.writeValueAsString(obj);
         } catch (JsonProcessingException e) {
